requirements.txt
# python3 -m pip install -r <file>    -r：一次安装多个模块
redis
records
pithy
pithy-test
pytest
prettyprinter
requests
pymysql
# Mysqlclient
allure-pytest
pytest-html
SQLAlchemy
xlrd
xlwt
PyYAML
nacos-sdk-python

# pithy --extra-index-url http://pypi.dev.elenet.me/eleme/eleme --trusted-host pypi.dev.elenet.me





# 生成报告命令：
# pytest -v test_suites --alluredir=rpt/xml --clean-alluredir    (或 test_suites/test_djh.py，也可不指定目录或文件 运行所有用例)
# allure generate rpt/xml -o rpt/html --clean

# pytest自带的报告：
# pytest --html=report.html --self-contained-html





